+++ Front-End/src/utilities/NodeUtils.tsx
@@ -1,7 +1,7 @@
 import { Edge, Node } from "react-flow-renderer"
 import CustomNodeData from "../visualizer/nodeMap/CustomNodeData"
 import NodeData from "./NodeData"
-import { toPng, toJpeg, toSvg } from "html-to-image";
+import { toPng, toJpeg } from "html-to-image";
 import download from "downloadjs";
 
 class NodeUtils {
